+++ src/lexer/lexer.c
@@ -19,7 +19,7 @@
 
 int	checkfile(char *name)
 {
-	int	i;
+	size_t	i;
 
 	i = ft_strlen(name);
 	if (i < 4 || name[i - 1] != 't' || name[i - 2] != 'r' || name[i - 3] != '.')
@@ -80,7 +80,7 @@ char	**loader(char *path, int *errnum, int *size)
 	return (buff);
 }
 
-//cant handle trailing whitespaces before nl
+//can't handle trailing whitespaces before nl
 int	lexer(char *path, t_mixer *m_data, int *errnum)
 {
 	char	**buffer;
